v3/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
orjson
qrcode[pil]==7.4.2
Pillow==10.4.0
qrcode[pil]==7.4.2
Pillow==10.4.0
